Game Progress refactor

+++ ShootingRangeMiniGame/Assets/Data/DataProvider.cs
@@ -5,8 +5,8 @@ namespace ShootingRangeMiniGame.Assets.Data
 	public class DataProvider
 	{
 		public int TargetsCount { get; private set; }
-		public float ProjectileSpeed { get; private set; }
-		public float Time { get; private set; }
+		public int ProjectileSpeed { get; private set; }
+		public int Time { get; private set; }
 		public int BulletsCount { get; private set; }
 		
 		public DataProvider()
@@ -25,8 +25,8 @@ namespace ShootingRangeMiniGame.Assets.Data
 				MatchCollection matches = Regex.Matches(data, pattern);
 				
 				TargetsCount = int.Parse(matches[0].Value);
-				ProjectileSpeed = float.Parse(matches[1].Value);
-				Time = float.Parse(matches[2].Value);
+				ProjectileSpeed = int.Parse(matches[1].Value);
+				Time = int.Parse(matches[2].Value);
 				BulletsCount = int.Parse(matches[3].Value);
 				
 				streamReader.Close();
